+++ lib/collision/object_collision.dart
@@ -2,7 +2,7 @@ import 'dart:ui';
 
 import 'package:bonfire/bonfire.dart';
 import 'package:bonfire/collision/collision_config.dart';
-import 'package:bonfire/util/extensions.dart';
+import 'package:bonfire/util/extensions/extensions.dart';
 import 'package:bonfire/util/vector2rect.dart';
 import 'package:flutter/material.dart';