+++ lib/geometry/rectangle.dart
@@ -1,7 +1,7 @@
 import 'dart:ui';
 
 import 'package:bonfire/geometry/shape.dart';
-import 'package:bonfire/util/extensions.dart';
+import 'package:bonfire/util/extensions/extensions.dart';
 import 'package:flame/extensions.dart';
 
 class RectangleShape extends Shape {