+++ lib/objects/sprite_object.dart
@@ -1,7 +1,7 @@
 import 'dart:ui';
 
 import 'package:bonfire/base/game_component.dart';
-import 'package:bonfire/util/extensions.dart';
+import 'package:bonfire/util/extensions/extensions.dart';
 import 'package:flame/sprite.dart';
 
 class SpriteObject extends GameComponent {